requirements.txt
altgraph==0.17
asgiref==3.2.10
click==7.1.2
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-datatables==0.5.2
itsdangerous==1.1.0
macholib==1.14
MarkupSafe==1.1.1
pdfrw==0.4
Pillow==8.3.1
python-dateutil==2.8.1
pytz==2020.1
qrcode==7.2
six==1.15.0
sqlparse==0.3.1
tinycss2==1.1.0
WeasyPrint==0.42.3
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
xlrd==1.2.0